// src/app/pages/models/user.ts
export class User {
  id: number;
  firstName: string;
  lastName: string;
  email: number;
  secondEmail: string;
  phone: string;
  secondPhone :string;
  address: string;
  matriculateFiscal: string;
  company: string;
  country : string;
  city: string;
  postalCode: number;
  webSiteLink : string;
  linkedinUrl: string;
  status : string ;
  archived : boolean;
}